Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

Base.Matrix

LinearOperators.LBFGSData

LinearOperators.LBFGSOperator

LinearOperators.LBFGSOperator

LinearOperators.LSR1Data

LinearOperators.LSR1Operator

LinearOperators.LSR1Operator

LinearOperators.LinearOperator

LinearOperators.LinearOperator

LinearOperators.LinearOperator

LinearOperators.LinearOperator

LinearOperators.LinearOperator

LinearOperators.LinearOperator

LinearOperators.LinearOperator

LinearOperators.PreallocatedLinearOperator

LinearOperators.PreallocatedLinearOperator

LinearOperators.PreallocatedLinearOperator

LinearOperators.PreallocatedLinearOperator

LinearOperators.PreallocatedLinearOperator

LinearOperators.PreallocatedLinearOperator

LinearOperators.TimedLinearOperator

LinearOperators.opEye

LinearOperators.opEye

LinearOperators.opEye

Base.kron

Base.push!

Base.push!

Base.show

Base.show

Base.size

Base.size

LinearAlgebra.diag

LinearAlgebra.diag

LinearOperators.BlockDiagonalOperator

LinearOperators.InverseLBFGSOperator

LinearOperators.check_ctranspose

LinearOperators.check_hermitian

LinearOperators.check_positive_definite

LinearOperators.hermitian

LinearOperators.opCholesky

LinearOperators.opDiagonal

LinearOperators.opDiagonal

LinearOperators.opExtension

LinearOperators.opHermitian

LinearOperators.opHermitian

LinearOperators.opHouseholder

LinearOperators.opInverse

LinearOperators.opLDL

LinearOperators.opOnes

LinearOperators.opRestriction

LinearOperators.opZeros

LinearOperators.reset!

LinearOperators.reset!

LinearOperators.reset!

LinearOperators.reset!

LinearOperators.reset!

LinearOperators.shape

LinearOperators.symmetric

​

Base.Matrix — Method

A = Matrix(op)

Materialize an operator as a dense array using op.ncol products.

LinearOperators.LBFGSData — Type

A data type to hold information relative to LBFGS operators.

LinearOperators.LBFGSOperator — Type

A type for limited-memory BFGS approximations.

LinearOperators.LBFGSOperator — Method

LBFGSOperator(T, n; [mem=5, scaling=true])
LBFGSOperator(n; [mem=5, scaling=true])

Construct a limited-memory BFGS approximation in forward form. If the type T is omitted, then Float64 is used.

LinearOperators.LSR1Data — Type

A data type to hold information relative to LSR1 operators.

LinearOperators.LSR1Operator — Type

A type for limited-memory SR1 approximations.

LinearOperators.LSR1Operator — Method

LSR1Operator(T, n; [mem=5, scaling=false)
LSR1Operator(n; [mem=5, scaling=false)

Construct a limited-memory SR1 approximation in forward form. If the type T is omitted, then Float64 is used.

LinearOperators.LinearOperator — Type

Base type to represent a linear operator. The usual arithmetic operations may be applied to operators to combine or otherwise alter them. They can be combined with other operators, with matrices and with scalars. Operators may be transposed and conjugate-transposed using the usual Julia syntax.

LinearOperators.LinearOperator — Type

LinearOperator(nrow, ncol, symmetric, hermitian, prod,
                [tprod=nothing,
                ctprod=nothing])

Construct a linear operator from functions.

LinearOperators.LinearOperator — Method

LinearOperator(M; symmetric=false, hermitian=false)

Construct a linear operator from a dense or sparse matrix. Use the optional keyword arguments to indicate whether the operator is symmetric and/or hermitian.

LinearOperators.LinearOperator — Method

LinearOperator(M)

Constructs a linear operator from a Hermitian matrix. If its elements are real, it is also symmetric.

LinearOperators.LinearOperator — Method

LinearOperator(M)

Constructs a linear operator from a symmetric tridiagonal matrix. If its elements are real, it is also Hermitian, otherwise complex symmetric.

LinearOperators.LinearOperator — Method

LinearOperator(M)

Constructs a linear operator from a symmetric matrix. If its elements are real, it is also Hermitian, otherwise complex symmetric.

LinearOperators.LinearOperator — Method

LinearOperator(type, nrow, ncol, symmetric, hermitian, prod,
                [tprod=nothing,
                ctprod=nothing])

Construct a linear operator from functions where the type is specified as the first argument. Notice that the linear operator does not enforce the type, so using a wrong type can result in errors. For instance,

A = [im 1.0; 0.0 1.0] # Complex matrix
op = LinearOperator(Float64, 2, 2, false, false, v->A*v, u->transpose(A)*u, w->A'*w)
Matrix(op) # InexactError

The error is caused because Matrix(op) tries to create a Float64 matrix with the contents of the complex matrix A.

LinearOperators.PreallocatedLinearOperator — Type

Type to represent a linear operator with preallocation. Implicit modifications may happen if used without care:

op = PreallocatedLinearOperator(rand(5, 5))
v  = rand(5)
x = op * v        # Uses internal storage and passes pointer to x
y = op * ones(5)  # Overwrites the same memory as x.
y === x           # true. op * v is lost

x = op * v        # Uses internal storage and passes pointer to x
y = op * x        # Silently overwrite x to zeros! Equivalent to mul!(x, A, x).
y == zeros(5)     # true. op * v and op * x are lost

LinearOperators.PreallocatedLinearOperator — Method

PreallocatedLinearOperator(M)

Constructs a linear operator from a Hermitian matrix. If its elements are real, it is also symmetric.

LinearOperators.PreallocatedLinearOperator — Method

PreallocatedLinearOperator(M)

Constructs a linear operator from a symmetric tridiagonal matrix. If its elements are real, it is also Hermitian, otherwise complex symmetric.

LinearOperators.PreallocatedLinearOperator — Method

PreallocatedLinearOperator(M)

Constructs a linear operator from a symmetric matrix. If its elements are real, it is also Hermitian, otherwise complex symmetric.

LinearOperators.PreallocatedLinearOperator — Method

PreallocatedLinearOperator(Mv, Mtu, Maw, M; symmetric=false, hermitian=false)

Construct a linear operator from a dense or sparse matrix, using Mv as storage space for M * v and Mtu as storage space for transpose(M) * u and Maw to store adjoint(M) * w. Use the optional keyword arguments to indicate whether the operator is symmetric and/or hermitian.

LinearOperators.PreallocatedLinearOperator — Method

PreallocatedLinearOperator(Mv, M :: Symmetric{<:Real})

Construct a linear operator from a symmetric real square matrix M with preallocation using Mv as storage space.

LinearOperators.TimedLinearOperator — Method

TimedLinearOperator(op)

Creates a linear operator instrumented with timers from TimerOutputs.

LinearOperators.opEye — Type

opEye()

Identity operator.

opI = opEye()
v = rand(5)
@assert opI * v === v

LinearOperators.opEye — Method

opEye(T, nrow, ncol)
opEye(nrow, ncol)

Rectangular identity operator of size nrowxncol and of data type T (defaults to Float64).

LinearOperators.opEye — Method

opEye(T, n)
opEye(n)

Identity operator of order n and of data type T (defaults to Float64).

Base.kron — Method

kron(A, B)

Kronecker tensor product of A and B in linear operator form, if either or both are linear operators. If both A and B are matrices, then Base.kron is used.

Base.push! — Function

push!(op, s, y)
push!(op, s, y, α, g)

Push a new {s,y} pair into a L-BFGS operator. The second calling sequence is used in inverse LBFGS updating in conjunction with damping, where α is the most recent steplength and g the gradient used when solving d=-Hg. In forward updating with damping, it is not necessary to supply α and g.

Base.push! — Method

push!(op, s, y)

Push a new {s,y} pair into a L-SR1 operator.

Base.show — Method

show(io, op)

Display basic information about a linear operator.

Base.show — Method

show(io, op)

Display basic information about a linear operator.

Base.size — Method

m = size(op, d)

Return the size of a linear operator along dimension d.

Base.size — Method

m, n = size(op)

Return the size of a linear operator as a tuple.

LinearAlgebra.diag — Method

diag(op)
diag!(op, d)

Extract the diagonal of a L-BFGS operator in forward mode.

LinearAlgebra.diag — Method

diag(op)
diag!(op, d)

Extract the diagonal of a L-SR1 operator in forward mode.

LinearOperators.BlockDiagonalOperator — Method

BlockDiagonalOperator(M1, M2, ..., Mn)

Creates a block-diagonal linear operator:

[ M1           ]
[    M2        ]
[       ...    ]
[           Mn ]

LinearOperators.InverseLBFGSOperator — Method

InverseLBFGSOperator(T, n, [mem=5; scaling=true])
InverseLBFGSOperator(n, [mem=5; scaling=true])

Construct a limited-memory BFGS approximation in inverse form. If the type T is omitted, then Float64 is used.

LinearOperators.check_ctranspose — Method

check_ctranspose(op)

Cheap check that the operator and its conjugate transposed are related.

LinearOperators.check_hermitian — Method

check_hermitian(op)

Cheap check that the operator is Hermitian.

LinearOperators.check_positive_definite — Method

check_positive_definite(op; semi=false)

Cheap check that the operator is positive (semi-)definite.

LinearOperators.hermitian — Method

hermitian(op)
ishermitian(op)

Determine whether the operator is Hermitian.

LinearOperators.opCholesky — Method

opCholesky(M, [check=false])

Inverse of a Hermitian and positive definite matrix as a linear operator using its Cholesky factorization. The factorization is computed only once. The optional check argument will perform cheap hermicity and definiteness checks.

LinearOperators.opDiagonal — Method

opDiagonal(d)

Diagonal operator with the vector d on its main diagonal.

LinearOperators.opDiagonal — Method

opDiagonal(nrow, ncol, d)

Rectangular diagonal operator of size nrow-by-ncol with the vector d on its main diagonal.

LinearOperators.opExtension — Method

Z = opExtension(I, ncol)
Z = opExtension(:, ncol)

Creates a LinearOperator extending a vector of size length(I) to size ncol, where the position of the elements on the new vector are given by the indices I. The operation w = Z * v is equivalent to w = zeros(ncol); w[I] = v.

Z = opExtension(k, ncol)

Alias for opExtension([k], ncol).

LinearOperators.opHermitian — Method

opHermitian(A)

A symmetric/hermitian operator based on a matrix.

LinearOperators.opHermitian — Method

opHermitian(d, A)

A symmetric/hermitian operator based on the diagonal d and lower triangle of A.

LinearOperators.opHouseholder — Method

opHouseholder(h)

Apply a Householder transformation defined by the vector h. The result is x -> (I - 2 h h') x.

LinearOperators.opInverse — Method

opInverse(M; symmetric=false, hermitian=false)

Inverse of a matrix as a linear operator using \. Useful for triangular matrices. Note that each application of this operator applies \.

LinearOperators.opLDL — Method

opLDL(M, [check=false])

Inverse of a symmetric matrix as a linear operator using its LDL' factorization if it exists. The factorization is computed only once. The optional check argument will perform a cheap hermicity check.

LinearOperators.opOnes — Method

opOnes(T, nrow, ncol)
opOnes(nrow, ncol)

Operator of all ones of size nrow-by-ncol and of data type T (defaults to Float64).

LinearOperators.opRestriction — Method

Z = opRestriction(I, ncol)
Z = opRestriction(:, ncol)

Creates a LinearOperator restricting a ncol-sized vector to indices I. The operation Z * v is equivalent to v[I]. I can be :.

Z = opRestriction(k, ncol)

Alias for opRestriction([k], ncol).

LinearOperators.opZeros — Method

opZeros(T, nrow, ncol)
opZeros(nrow, ncol)

Zero operator of size nrow-by-ncol and of data type T (defaults to Float64).

LinearOperators.reset! — Method

reset!(op)

Reset the product counters of a linear operator.

LinearOperators.reset! — Method

reset!(op)

Resets the LBFGS data of the given operator.

LinearOperators.reset! — Method

reset!(op)

Resets the LSR1 data of the given operator.

LinearOperators.reset! — Method

reset!(data)

Reset the given LSR1 data.

LinearOperators.reset! — Method

reset!(data)

Resets the given LBFGS data.

LinearOperators.shape — Method

m, n = shape(op)

An alias for size.

LinearOperators.symmetric — Method

symmetric(op)
issymmetric(op)

Determine whether the operator is symmetric.